requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
tensorflow>=2.16.1
python-multipart>=0.0.7
numpy>=1.26.4
Pillow>=10.3.0
python-dotenv>=1.0.0
opencv-python>=4.5.0